userPortal(Flask,Svelte)/Tarragon-Solutions/backend/requirements.txt
Flask==1.1.2
Flask-Login==0.5.0
Flask-WTF==0.14.3
# Flask-MySQLdb==0.2.0
mysql-connector
# mysqlclient==1.4.6
# Flask-MySQL==1.5.1
# PyMySQL==0.10.0
